Schematics update, changes in hardware were made in order to fix electrical surge caused reboot issue.
1)Changed power system, connected alpha and beta (now there is no no diode 
between). That helps to keep LCD from rebooting (with diode, beta voltage was 
next to lower voltage limit of LCD). Diode separation was historical (another 
attempt to fix rebooting).
2)Added TWI anti-surge capacitors. A 33 nf capacitor together with 10k resistor
has relaxation time equivalent frequency of about 3000 Hz. TWI data rate is 
set to 770Hz, this shows that the capacitor value is big, but still not big 
enough to break TWI communications. 
Adding the two capacitors protects microcontroller from surges, generated by
turning heater on/off. The trouble is that both heater and DS1621 thermometer 
reside in the same aquarium, under water. There is a definite experimental 
evidence, that water helps surge propagation. 
In experiment, heater was plugged in and out many times. Without capacitors, 
that caused reboots, when thermometer was in water, and didn't cause reboots, 
when thermometer was out in the air. After adding 33nf capacitors, reboots 
stopped to occur when thermometer was in water.